Add files via upload
The initial commit from before had some problems so consider this the initial commit. There are still some problems and implementations that need to be done. For 1 the updates on the dev page arent working correctly. The cache isnt being properly cleared. Still need to implement an adavanced functon ( The advanced function is going to be implementing those things on the index.html page, the recommendations and last inputted in the DB).