Add files via upload
added updates. Fixed book_id searches as they were not working, implemented the navigation tab. Browse Books just links to a search of all books in the database. categories now has a dropdown that does a search of all the books in a specific category. Tried implementing a recommended function that works by using past searches. not working correctly but it is doing something that can be built upon. I added a back to home button in bottom of the results page. changed some fonts and sizes to look better. Updates work on title only, still cannot update stuff like bookID, authorname, or authorID.